Refactor: type 분리
