+++ src/reducer/context.tsx
@@ -1,22 +1,6 @@
-import React, { useContext, useState, createContext, useReducer, Dispatch } from 'react';
+import React, { useContext, createContext, useReducer, Dispatch } from 'react';
 import reducer from './reducer';
-
-export enum DataActionTypes {
-  UPDATE_TPS_DATA = 'UPDATE_TPS_DATA',
-}
-
-export type State = {
-  tps: ITps;
-};
-
-export type Action = { type: DataActionTypes.UPDATE_TPS_DATA; data: number };
-
-type DataDispatch = Dispatch<Action>;
-
-interface ITps {
-  data: number[];
-  error?: string;
-}
+import { State, DataDispatch } from './types';
 
 const INITIAL_STATE = {
   tps: {
